fix: preserve fresh identity across bootstrap retries
