+++ backend/tests/unit/test_runtime_bootstrap_tasks.py
@@ -4,8 +4,6 @@ from __future__ import annotations
 
 import pytest
 
-from app.domain.relative_strength import BALANCED_RS_FORMULA_VERSION
-
 
 @pytest.fixture(autouse=True)
 def classify_unit_test_bootstraps_as_non_fresh(monkeypatch):
@@ -467,94 +465,6 @@ def test_queue_local_runtime_bootstrap_splits_primary_and_background_market_chai
             "queue_state": "queued",
         },
     ]
-
-
-@pytest.mark.parametrize("fresh_install", [True, False])
-def test_queue_bootstrap_captures_pristine_installation_once(
-    monkeypatch,
-    fresh_install,
-):
-    from app.tasks import runtime_bootstrap_tasks as module
-
-    class _FakeAsyncResult:
-        def __init__(self, task_id: str) -> None:
-            self.id = task_id
-
-    classifications = []
-    saved = []
-    queued_operations = []
-    completion_payloads = []
-
-    def _classify():
-        classifications.append(fresh_install)
-        return fresh_install
-
-    monkeypatch.setattr(module, "_is_fresh_install_at_dispatch", _classify)
-    monkeypatch.setattr(
-        module,
-        "record_runtime_bootstrap_run",
-        lambda **payload: saved.append(payload) or payload,
-    )
-    def _queue(market_plan, **_kwargs):
-        queued_operations.append(
-            [stage.operation for stage in market_plan.stages]
-        )
-        completion_payloads.append(dict(_kwargs["completion_kwargs"]))
-        return _FakeAsyncResult(f"task-{market_plan.market.lower()}")
-
-    monkeypatch.setattr(module, "_queue_market_bootstrap_workflow", _queue)
-
-    module.queue_local_runtime_bootstrap(
-        primary_market="US",
-        enabled_markets=("US", "HK"),
-    )
-
-    assert classifications == [fresh_install]
-    assert saved
-    assert {record["fresh_install"] for record in saved} == {fresh_install}
-    expected_operation = (
-        module.BootstrapOperation.BOOTSTRAP_BALANCED_MARKET_RS
-        if fresh_install
-        else module.BootstrapOperation.CALCULATE_MARKET_RS_SNAPSHOT
-    )
-    assert queued_operations
-    assert all(expected_operation in operations for operations in queued_operations)
-    if fresh_install:
-        assert all(
-            payload["expected_formula_version"]
-            == BALANCED_RS_FORMULA_VERSION
-            for payload in completion_payloads
-        )
-    else:
-        assert all(
-            "expected_formula_version" not in payload
-            for payload in completion_payloads
-        )
-
-
-def test_fresh_bootstrap_signature_routes_activation_to_market_queue() -> None:
-    from app.domain.bootstrap.plan import build_bootstrap_plan
-    from app.tasks.runtime_bootstrap_tasks import _build_market_bootstrap_signatures
-
-    market_plan = build_bootstrap_plan(
-        primary_market="HK",
-        enabled_markets=("HK",),
-        fresh_install=True,
-    ).market_plans[0]
-
-    signatures = _build_market_bootstrap_signatures(market_plan)
-    activation = next(
-        signature
-        for signature in signatures
-        if signature.task
-        == "app.tasks.market_rs_tasks.bootstrap_balanced_market_rs"
-    )
-
-    assert activation.kwargs == {
-        "market": "HK",
-        "activity_lifecycle": "bootstrap",
-    }
-    assert activation.options["queue"] == "market_jobs_hk"
 
 
 def test_queue_local_runtime_bootstrap_does_not_dispatch_when_initial_manifest_fails(monkeypatch):
@@ -1069,107 +979,3 @@ def test_complete_background_market_bootstrap_marks_market_failure_without_globa
             "message": "Bootstrap scan did not publish",
         }
     ]
-
-
-@pytest.mark.parametrize(
-    ("completion_task_name", "task_kwargs", "expected_result"),
-    [
-        (
-            "complete_local_runtime_bootstrap",
-            {"primary_market": "US"},
-            {
-                "status": "failed",
-                "primary_market": "US",
-                "market": "US",
-                "reason": "balanced market rs formula not active",
-            },
-        ),
-        (
-            "complete_background_market_bootstrap",
-            {"market": "US"},
-            {
-                "status": "failed",
-                "market": "US",
-                "reason": "balanced market rs formula not active",
-            },
-        ),
-    ],
-)
-def test_fresh_bootstrap_completion_rejects_legacy_formula_pointer(
-    monkeypatch,
-    completion_task_name,
-    task_kwargs,
-    expected_result,
-):
-    from app.services.bootstrap_readiness_service import (
-        BootstrapReadiness,
-        MarketBootstrapReadiness,
-    )
-    from app.tasks import runtime_bootstrap_tasks as module
-
-    class _FakeSession:
-        def close(self):
-            pass
-
-    class _FakeReadinessService:
-        def evaluate(
-            self,
-            db,
-            *,
-            enabled_markets,
-            bootstrap_started_at=None,
-            expected_formula_versions=None,
-        ):
-            calls["expectations"] = expected_formula_versions
-            return BootstrapReadiness(
-                empty_system=False,
-                market_results={
-                    "US": MarketBootstrapReadiness(
-                        market="US",
-                        core_ready=True,
-                        scan_ready=True,
-                        rs_ready=False,
-                    )
-                },
-            )
-
-    calls = {}
-    bootstrap_states = []
-    failed_markets = []
-    monkeypatch.setattr(module, "SessionLocal", lambda: _FakeSession())
-    monkeypatch.setattr(
-        "app.services.bootstrap_readiness_service.BootstrapReadinessService",
-        _FakeReadinessService,
-    )
-    monkeypatch.setattr(
-        "app.services.runtime_preferences_service.get_runtime_preferences",
-        lambda _db: type("Prefs", (), {"bootstrap_started_at": None})(),
-    )
-    monkeypatch.setattr(
-        "app.services.runtime_preferences_service.set_bootstrap_state",
-        lambda _db, state: bootstrap_states.append(state),
-    )
-    monkeypatch.setattr(
-        module,
-        "mark_market_activity_failed",
-        lambda _db, **kwargs: failed_markets.append(kwargs),
-    )
-
-    task = getattr(module, completion_task_name)
-    result = task.run(
-        **task_kwargs,
-        expected_formula_version=BALANCED_RS_FORMULA_VERSION,
-    )
-
-    assert result == expected_result
-    assert calls["expectations"] == {
-        "US": BALANCED_RS_FORMULA_VERSION,
-    }
-    assert failed_markets[0]["stage_key"] == "market_rs"
-    assert failed_markets[0]["message"] == (
-        "Balanced Market RS activation incomplete"
-    )
-    if completion_task_name == "complete_local_runtime_bootstrap":
-        assert bootstrap_states == ["failed"]
-    else:
-        assert bootstrap_states == []
